+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/views.cljs
@@ -87,17 +87,18 @@
     (when is-add?
       (dispatch [::events/get-ssh-keys-not-associated #(reset! ssh-keys %)]))
     (fn [_operation on-change-fn]
-      [ui/FormDropdown
-       {:label       "SSH key"
-        :loading     (nil? @ssh-keys)
-        :placeholder (if is-add?
-                       (@tr [:leave-empty-to-generate-ssh-keypair])
-                       (@tr [:select-credential]))
-        :on-change   (ui-callback/value on-change-fn)
-        :clearable   true
-        :options     (map (fn [{:keys [id name]}]
-                            {:key id, :text (or name id), :value id}) @ssh-keys)
-        :selection   true}])))
+      (if (nil? @ssh-keys)
+        [ui/Message {:info true} (@tr [:no-credentials-to-remove])]
+        [ui/FormDropdown
+         {:label       "SSH key"
+          :placeholder (if is-add?
+                         (@tr [:leave-empty-to-generate-ssh-keypair])
+                         (@tr [:select-credential]))
+          :on-change   (ui-callback/value on-change-fn)
+          :clearable   true
+          :options     (map (fn [{:keys [id name]}]
+                              {:key id, :text (or name id), :value id}) @ssh-keys)
+          :selection   true}]))))
 
 
 (defn DropdownReleases
@@ -117,7 +118,7 @@
   [{:keys [id] :as _resource} operation show? _title _icon _button-text]
   (let [tr            (subscribe [::i18n-subs/tr])
         close-fn      #(reset! show? false)
-        form-data     (r/atom {:execution-mode "push"})
+        form-data     (r/atom {})
         key-data      (r/atom nil)
         loading?      (r/atom nil)
         on-change-fn  (fn [k v]
@@ -135,47 +136,39 @@
                          (dispatch [::events/operation id operation @form-data
                                     on-success-fn on-error-fn]))]
     (fn [_resource operation show? title icon button-text]
-      [ui/Modal
-       {:open       @show?
-        :close-icon true
-        :on-close   close-fn
-        :trigger    (r/as-element
-                      [ui/MenuItem {:on-click #(reset! show? true)}
-                       [ui/Icon {:name icon}]
-                       title])}
-       [uix/ModalHeader {:header title}]
-       [ui/ModalContent
-        [ui/Form
-         [ui/FormDropdown {:label         "Execution mode"
-                           :selection     true
-                           :default-value (:execution-mode @form-data)
-                           :on-change     (ui-callback/value (partial on-change-fn :execution-mode))
-                           :options       [{:key "push", :text "push", :value "push"}
-                                           {:key "mixed", :text "mixed", :value "mixed"}
-                                           {:key "pull", :text "pull", :value "pull"}]}]
-         [SshKeysDropdown operation (partial on-change-fn :credential)]
+      (let [credential (:credential @form-data)]
+        [ui/Modal
+         {:open       @show?
+          :close-icon true
+          :on-close   close-fn
+          :trigger    (r/as-element
+                        [ui/MenuItem {:on-click #(reset! show? true)}
+                         [ui/Icon {:name icon}]
+                         title])}
+         [uix/ModalHeader {:header title}]
+         [ui/ModalContent
+          [ui/Form
+           [SshKeysDropdown operation (partial on-change-fn :credential)]
 
-         (when @key-data
-           [ui/FormField
-            [:a {:href     (str "data:text/plain;charset=utf-8,"
-                                (js/encodeURIComponent @key-data))
-                 :target   "_blank"
-                 :download "ssh_private.key"}
-             [ui/Button {:positive       true
-                         :fluid          true
-                         :icon           "download"
-                         :label-position "left"
-                         :as             "div"
-                         :content        (@tr [:download])}]]])]]
-       [ui/ModalActions
-        [uix/Button
-         {:text     (@tr [:cancel])
-          :on-click close-fn}]
-        [uix/Button
-         {:text     (if @key-data (@tr [:close]) button-text)
-          :primary  true
-          :loading  (true? @loading?)
-          :on-click (if @key-data close-fn on-click-fn)}]]])))
+           (when @key-data
+             [ui/FormField
+              [:a {:href     (str "data:text/plain;charset=utf-8,"
+                                  (js/encodeURIComponent @key-data))
+                   :target   "_blank"
+                   :download "ssh_private.key"}
+               [ui/Button {:positive       true
+                           :fluid          true
+                           :icon           "download"
+                           :label-position "left"
+                           :as             "div"
+                           :content        (@tr [:download])}]]])]]
+         [ui/ModalActions
+          [uix/Button
+           {:text     (if @key-data (@tr [:close]) button-text)
+            :primary  true
+            :disabled (nil? credential)
+            :loading  (true? @loading?)
+            :on-click (if @key-data close-fn on-click-fn)}]]]))))
 
 
 (defn is-old-version?
@@ -196,10 +189,12 @@
         releases      (subscribe [::edge-subs/nuvlabox-releases-options])
         close-fn      #(reset! show? false)
         form-data     (r/atom nil)
+        force-restart (r/atom false)
         project       (-> @status :installation-parameters :project-name)
         working-dir   (-> @status :installation-parameters :working-dir)
         config-files  (-> @status :installation-parameters :config-files)
         environment   (-> @status :installation-parameters :environment)
+        nb-version    (get @status :nuvlabox-engine-version nil)
         on-change-fn  #(swap! form-data assoc :nuvlabox-release %)
         on-success-fn close-fn
         on-error-fn   close-fn
@@ -211,9 +206,11 @@
     (swap! form-data assoc :working-dir working-dir)
     (swap! form-data assoc :config-files (str/join "\n" config-files))
     (swap! form-data assoc :environment (str/join "\n" environment))
+    (swap! form-data assoc :force-restart false)
+    (when nb-version
+      (swap! form-data assoc :current-version nb-version))
     (fn [{:keys [id] :as _resource} _operation show? title icon button-text]
       (let [correct-nb?    (= (:parent @status) id)
-            nb-version     (get @status :nuvlabox-engine-version "")
             target-version (->> @releases
                                 (some #(when (= (:value %) (:nuvlabox-release @form-data)) %))
                                 :key)]
@@ -261,6 +258,18 @@
           [uix/Accordion
            [:<>
             [ui/Form
+             [ui/FormField
+              [:label
+               "Force Restart"]
+              [ui/Radio {:toggle    true
+                         :checked   @force-restart
+                         :label     (if (:force-restart @form-data)
+                                      (@tr [:nuvlabox-update-force-restart])
+                                      (@tr [:nuvlabox-update-no-force-restart]))
+                         :on-change #(do
+                                       (swap! force-restart not)
+                                       (swap! form-data assoc :force-restart @force-restart))}]]
+
              [ui/FormInput {:label         (str/capitalize (@tr [:project]))
                             :placeholder   "nuvlabox"
                             :required      true
@@ -292,9 +301,6 @@
            :title-size :h4
            :default-open false]]
          [ui/ModalActions
-          [uix/Button
-           {:text     (@tr [:cancel])
-            :on-click close-fn}]
           [uix/Button
            {:text     button-text
             :disabled (or (utils/form-update-data-incomplete? @form-data) (is-old-version? nb-version))
@@ -356,6 +362,7 @@
             is-join-manager?    (= cluster-action "join-manager")
             is-join-worker?     (= cluster-action "join-worker")
             is-leave-action?    (= cluster-action "leave")
+            is-new-action?      (= cluster-action "force-new-cluster")
             is-join-action?     (or is-join-worker? is-join-manager?)
             manager-selected?   (not (nil? nuvlabox-manager-id))
             valid-manager?      (and
@@ -388,6 +395,8 @@
                                     (swap! form-data dissoc :cluster-action)
                                     (do
                                       (swap! form-data dissoc :token)
+                                      (swap! form-data dissoc :advertise-addr)
+                                      (swap! form-data dissoc :nuvlabox-manager-status)
                                       (swap! form-data assoc :cluster-action value)
                                       ))
                                   (swap! form-data dissoc :nuvlabox-manager-id)))
@@ -423,7 +432,16 @@
                     [:p {:style {:font-weight "bold"}} (:token @join-token)]]])])
              (when is-leave-action?
                [ui/Message {:negative true
-                            :content  (@tr [:cluster-action-leave-warning])}])])]
+                            :content  (@tr [:cluster-action-leave-warning])}])
+             (when is-new-action?
+               [ui/FormInput
+                {:label       "Advertise address (optional)"
+                 :placeholder "<ip|interface>[:port]"
+                 :on-change   (ui-callback/value
+                                (fn [value]
+                                  (if (str/blank? value)
+                                    (swap! form-data dissoc :advertise-addr)
+                                    (swap! form-data assoc :advertise-addr value))))}])])]
 
          [ui/ModalActions
           [uix/Button