+++ code/src/cljs/sixsq/nuvla/ui/infrastructures/utils.cljs
@@ -20,6 +20,8 @@
 
 (def infra-service-subtype-exoscale "infrastructure-service-exoscale")
 (def infra-service-subtype-google "infrastructure-service-google")
+(def infra-service-subtype-openstack "infrastructure-service-openstack")
+
 ;; TODO: This needs to come from the server as configuration defaults.
 ;; Probably from template(s) or better a configuration resource?
 (def cloud-params-defaults
@@ -30,6 +32,10 @@
     :cloud-vm-image "Linux Ubuntu 18.04 LTS 64-bit"
     :cloud-security-group "docker-machine"
     :cloud-doc-link "https://www.exoscale.com/compute"}
+   "infrastructure-service-openstack"
+   {:cloud-vm-size "eo1.small"
+    :cloud-vm-disk-size 50
+    :cloud-doc-link "https://docs.openstack.org/wallaby/user/"}
    "infrastructure-service-amazonec2"
    {:cloud-vm-size "t2.medium"
     :cloud-vm-disk-size 50
@@ -101,22 +107,32 @@
         cloud-region          (get-in db [::spec/infra-service :cloud-region] (cloud-param-default-value mgmt-cred-subtype :cloud-region))
         cloud-project          (get-in db [::spec/infra-service :cloud-project] (cloud-param-default-value mgmt-cred-subtype :cloud-project))
         cloud-security-group  (get-in db [::spec/infra-service :cloud-security-group])
+        cloud-network         (get-in db [::spec/infra-service :cloud-network])
+        cloud-domain          (get-in db [::spec/infra-service :cloud-domain])
+        cloud-api-endpoint    (get-in db [::spec/infra-service :cloud-api-endpoint])
+        cloud-floating-ip     (get-in db [::spec/infra-service :cloud-floating-ip])
+        cloud-user         (get-in db [::spec/infra-service :cloud-user])
         acl                   (get-in db [::spec/infra-service :acl])]
     (-> {}
-        (assoc-in [:template :href] (str "infrastructure-service-template/" template-type))
-        (assoc-in [:template :name] service-name)
-        (assoc-in [:template :description] description)
-        (assoc-in [:template :parent] group-id)
-        (assoc-in [:template :subtype] subtype)
-        (cond-> acl (assoc-in [:template :acl] acl))
-        (cond-> endpoint (assoc-in [:template :endpoint] endpoint))
-        (cond-> (and (= template-type "coe") multiplicity) (assoc-in [:template :cluster-params :multiplicity] multiplicity))
-        (cond-> (and (= template-type "coe") (not-empty ssh-keys)) (assoc-in [:template :cluster-params :ssh-keys] ssh-keys))
-        (cond-> (and (= template-type "coe") (not-empty cloud-vm-image)) (assoc-in [:template :cluster-params :cloud-vm-image] cloud-vm-image))
-        (cond-> (and (= template-type "coe") (not-empty cloud-vm-size)) (assoc-in [:template :cluster-params :cloud-vm-size] cloud-vm-size))
-        (cond-> (and (= template-type "coe") cloud-vm-disk-size) (assoc-in [:template :cluster-params :cloud-vm-disk-size] cloud-vm-disk-size))
-        (cond-> (and (= template-type "coe") (not-empty cloud-region)) (assoc-in [:template :cluster-params :cloud-region] cloud-region))
-        (cond-> (and (= template-type "coe") (not-empty cloud-project)) (assoc-in [:template :cluster-params :cloud-project] cloud-project))
-        (cond-> (and (= template-type "coe") (not-empty cloud-security-group)) (assoc-in [:template :cluster-params :cloud-security-group] cloud-security-group))
-        (cond-> (and (= template-type "coe") coe-manager-install) (assoc-in [:template :cluster-params :coe-manager-install] coe-manager-install))
-        (cond-> (and (= template-type "coe") management-credential) (assoc-in [:template :management-credential] management-credential)))))
+      (assoc-in [:template :href] (str "infrastructure-service-template/" template-type))
+      (assoc-in [:template :name] service-name)
+      (assoc-in [:template :description] description)
+      (assoc-in [:template :parent] group-id)
+      (assoc-in [:template :subtype] subtype)
+      (cond-> acl (assoc-in [:template :acl] acl))
+      (cond-> endpoint (assoc-in [:template :endpoint] endpoint))
+      (cond-> (and (= template-type "coe") multiplicity) (assoc-in [:template :cluster-params :multiplicity] multiplicity))
+      (cond-> (and (= template-type "coe") (not-empty ssh-keys)) (assoc-in [:template :cluster-params :ssh-keys] ssh-keys))
+      (cond-> (and (= template-type "coe") (not-empty cloud-vm-image)) (assoc-in [:template :cluster-params :cloud-vm-image] cloud-vm-image))
+      (cond-> (and (= template-type "coe") (not-empty cloud-vm-size)) (assoc-in [:template :cluster-params :cloud-vm-size] cloud-vm-size))
+      (cond-> (and (= template-type "coe") cloud-vm-disk-size) (assoc-in [:template :cluster-params :cloud-vm-disk-size] cloud-vm-disk-size))
+      (cond-> (and (= template-type "coe") (not-empty cloud-region)) (assoc-in [:template :cluster-params :cloud-region] cloud-region))
+      (cond-> (and (= template-type "coe") (not-empty cloud-project)) (assoc-in [:template :cluster-params :cloud-project] cloud-project))
+      (cond-> (and (= template-type "coe") (not-empty cloud-security-group)) (assoc-in [:template :cluster-params :cloud-security-group] cloud-security-group))
+      (cond-> (and (= template-type "coe") (not-empty cloud-network)) (assoc-in [:template :cluster-params :cloud-network] cloud-network))
+      (cond-> (and (= template-type "coe") (not-empty cloud-domain)) (assoc-in [:template :cluster-params :cloud-domain] cloud-domain))
+      (cond-> (and (= template-type "coe") (not-empty cloud-api-endpoint)) (assoc-in [:template :cluster-params :cloud-api-endpoint] cloud-api-endpoint))
+      (cond-> (and (= template-type "coe") (not-empty cloud-floating-ip)) (assoc-in [:template :cluster-params :cloud-floating-ip] cloud-floating-ip))
+      (cond-> (and (= template-type "coe") (not-empty cloud-user)) (assoc-in [:template :cluster-params :cloud-user] cloud-user))
+      (cond-> (and (= template-type "coe") coe-manager-install) (assoc-in [:template :cluster-params :coe-manager-install] coe-manager-install))
+      (cond-> (and (= template-type "coe") management-credential) (assoc-in [:template :management-credential] management-credential)))))